--- js/basic-data.js
@@ -1,15 +1,0 @@
-import {getListElements} from './util.js';
-
-const MESSAGE_ARRAY = ['Всё отлично!', 'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-const NAME_ARRAY = ['Александр', 'Дмитрий', 'Пётр', 'Анна', 'Мария', 'Константин', 'Ирина', 'Анастасия'];
-const DESCRIPTION_NUMBER = 25;
-
-const ID_ARRAY = getListElements(25);
-const AVATAR_ARRAY = getListElements(6);
-
-export {MESSAGE_ARRAY, NAME_ARRAY, DESCRIPTION_NUMBER, ID_ARRAY, AVATAR_ARRAY};
